# escuagest-api/src/main/resources/static/openapi.yml
openapi: 3.0.2
info:
  title: Escuagest API
  description: "Escuagest API"
  contact:
    email: javier.ote.ote@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1-SNAPSHOT
servers:
  - url: http://localhost:8080
tags:
  - name: user
    description: Operations about user
    externalDocs:
      description: Find out more about our store
      url: http://swagger.io
paths:
  /user/{id}:
    get:
      tags:
        - user
      summary: Get user by id
      operationId: getUserById
      parameters:
        - name: id
          in: path
          description: 'The id that needs to be fetched'
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        404:
          description: User not found
          content: { }
  /user:
    get:
      tags:
        - user
      summary: Get all users
      operationId: getAllUsers
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        404:
          description: User not found
          content: { }
    post:
      tags:
        - user
      summary: Create a new user
      operationId: postNewUser
      requestBody:
        description: User object thats needs to be added on the system.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        201:
          description: User created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'


components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    InternalServerError:
      description: "Generic Internal Error"
      content:
        application/json:
          schema:
            $ref: '#components/schemas/InternalServerError'
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        fullname:
          type: string
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        fullname:
          type: string
      required:
        - username
        - fullname
    InternalServerError:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string

security:
  - bearerAuth: []